Move `persistent_mode` test to `sync` module
Group it because it starts a live mainnet peer.
